+++ ak_library_management/models/res_partner.py
@@ -10,4 +10,4 @@ class ResPartner(models.Model):
 
     not_trust_worthy = fields.Boolean(string="Not Trustworthy", help="Indicates if the customer is trustworthy or not.")
     is_member = fields.Boolean(string="Is Member", help="Indicates whether the customer is a member.")
-    slug = fields.Char(string='Slug', store=True)
+    # slug = fields.Char(string='Slug', store=True)